refactor(select)!: change the type of the value property to an array
The value property is always a list. Even when the multiple attribute is not set.
This way we ensure a more consistent API.
To simplify it further the the internal tempvalue property was removed too.
The value will now always be updated directly

BREAKING CHANGE: The value property is not a string anymore but an array.