feat(button): add href attribute so that the component can be used as a link
refactor(button): remove  property

BREAKING CHANGE: The  property is removed. The underlying <button> or <a> element now takes
up the whole width of the host element. The width of the host element can be controlled with css
from the outside. The only exception are configurations that only show an icon and therefor should
be displayed with an aspect ratio of 1.